Complete rewrite of the module
There is no *cmd.yml* file alongside the main program anymore and
parsers configurations are now in a *_cmd.yml* file in a (sub)command
directory. Alongside this file, an *_anchors.yml* file can manage
anchors for the command (and childs commands). With theses files, for a
parser, there can be YAML files defining commands or directories for
subparsers (following the same pattern as just describe).

Logic of Python commands are still put in a *commands* directory, as
package following arborescence of commands and a library directory can
still contains some customizations (one Pythonfile with the same name of
the customization for types, actions are argcomplete completers).

CLI configuration directory, commands directory, library directory can
be set while initializing the module.